split utils up further
