+++ corebehrt/functional/utils.py
@@ -1,42 +1,11 @@
 """Random utils, should be structered later"""
 
 from dataclasses import asdict, fields
-from datetime import datetime
-from typing import List, Union
+from typing import List
 
-import dask.dataframe as dd
 import pandas as pd
 
 from corebehrt.modules.preparation.dataset import PatientData
-
-
-def get_background_length(patients: List[PatientData], vocabulary) -> int:
-    """Get the length of the background sentence, first SEP token included."""
-    background_tokens = set([v for k, v in vocabulary.items() if k.startswith("BG_")])
-    example_concepts = patients[0].concepts
-    background_length = len(set(example_concepts) & background_tokens)
-    return background_length + 2  # +2 for [CLS] and [SEP] tokens
-
-
-def get_abspos_from_origin_point(
-    timestamps: Union[pd.Series, List[datetime]], origin_point: datetime
-) -> Union[pd.Series, List[float]]:
-    """Get the absolute position in hours from the origin point"""
-    if isinstance(timestamps, dd.Series):  # len can cause issues with dask
-        return (timestamps - origin_point).dt.total_seconds() / 60 / 60
-    if isinstance(timestamps, pd.Series):
-        if len(timestamps) == 0:
-            return pd.Series([])
-        return (timestamps - origin_point).dt.total_seconds() / 60 / 60
-    elif isinstance(timestamps, list):
-        return [
-            (timestamp - origin_point).total_seconds() / 60 / 60
-            for timestamp in timestamps
-        ]
-    else:
-        raise TypeError(
-            "Invalid type for timestamps, only pd.Series and list are supported."
-        )
 
 
 def filter_table_by_pids(df: pd.DataFrame, pids: List[str]) -> pd.DataFrame:
@@ -53,17 +22,6 @@ def remove_missing_timestamps(df: pd.DataFrame) -> pd.DataFrame:
     Returns a new table with only the rows that have a TIMESTAMP.
     """
     return df[df.TIMESTAMP.notna()]
-
-
-def get_non_priority_tokens(vocabulary: dict, low_priority_prefixes: List[str]) -> set:
-    """
-    Get tokens that start with low_priority_prefixes.
-    """
-    return {
-        v
-        for k, v in vocabulary.items()
-        if any(k.startswith(prefix) for prefix in low_priority_prefixes)
-    }
 
 
 def subset_patient_data(patient: PatientData, keep_indices: List[int]) -> PatientData:
