+++ tests/test_modules/test_outcomemaker.py
@@ -3,7 +3,7 @@ import unittest
 import pandas as pd
 
 from corebehrt.modules.cohort_handling.outcomes import OutcomeMaker
-from corebehrt.functional.utils import get_abspos_from_origin_point
+from corebehrt.functional.preparation.utils import get_abspos_from_origin_point
 from datetime import datetime
